Rename exchanges.rs to sources.rs
